### 2025-09-05 #8945 @editorialbot
[PRE REVIEW]: TACtool: A targeting and co-ordination tool for spatial sampling and data traceability applied to scanning electron microscopy and laser ablation ICP-MS
**Submitting author:** @leorudczenko (<a href="http://orcid.org/0009-0005-8941-1387">Leo Rudczenko</a>)
**Repository:** https://github.com/britishgeologicalsurvey/tactool
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.3.0
