### 2025-09-08 4ec7ee7
Fix invalid DOIs
