+++ paper/paper.bib
@@ -43,7 +43,7 @@
     pages={464-477},
     year={2018},
     issn={0024-4937},
-    doi={https://doi.org/10.1016/j.lithos.2018.08.033},
+    doi={10.1016/j.lithos.2018.08.033},
     url={https://www.sciencedirect.com/science/article/pii/S0024493718303141},
     author={Rebecca L. Astbury and Maurizio Petrelli and Teresa Ubide and Michael J. Stock and Ilenia Arienzo and Massimo D'Antonio and Diego Perugini},
     keywords={Campi Flegrei, Astroni, LA-ICP-MS, Trace elements, Magma mixing, Eruption timescales},
@@ -57,7 +57,7 @@
     pages={113-132},
     year={2020},
     issn={0016-7037},
-    doi={https://doi.org/10.1016/j.gca.2020.05.019},
+    doi={10.1016/j.gca.2020.05.019},
     url={https://www.sciencedirect.com/science/article/pii/S0016703720303240},
     author={Christophe Bonnetti and Lingli Zhou and Thomas Riegler and Joël Brugger and Martin Fairclough},
     keywords={Roll front uranium deposits, Activity cycle, S isotope and trace element fractionation, Pyrite composition},
