### 2025-09-08 #8945 @leorudczenko
Hi @kthyng and thanks for handling our paper! I've just pushed a commit to fix the invalid DOIs picked up by the bot. I believe the following reviewers would be good for this paper based on their expertise:
- nvnsudharsan
- laurenguido
- dIOGOLOC
- hulecom
- shipengcheng1230
