### 2025-12-05 #8945 @dIOGOLOC
Hi @philipcardiff ,

Thank you very much for the invitation to review the submission “TACtool”. I appreciate being considered for this role. Unfortunately, I am unable to accept the invitation at this time, as I have very limited availability toward the end of the year.

Thank you again for the invitation.
