### 2025-12-10 #8945 @philipcardiff
Hello @IMBalENce,

Would you be interested in reviewing the submission “TACtool: A targeting and co-ordination tool for spatial sampling and data traceability applied to scanning electron microscopy and laser ablation ICP-MS ”, which we received at JOSS? The link for the submission is https://github.com/openjournals/joss-reviews/issues/8945, and the repository is at https://github.com/britishgeologicalsurvey/tactool.

If you are unfamiliar with JOSS, the review process is unique: it takes place in a GitHub issue, is open, and encourages author-reviewer-editor conversations. You can have a look at our review criteria (https://joss.readthedocs.io/en/latest/review_criteria.html) and review checklist (https://joss.readthedocs.io/en/latest/review_checklist.html) to get an idea of what is expected.
