### 2026-01-14 d3e6cad
Adjust paper structure to fit new JOSS requirements
