+++ paper/paper.md
@@ -34,7 +34,13 @@ bibliography: paper.bib
 
 The precise analysis of geological samples is crucial for understanding the history of Earths formation, identification of mineral deposits, age dating, paleoclimatology and more [@Chew:2021]. Laser ablation inductively coupled plasma mass spectrometry (LA-ICP-MS) has become one of the most common analytical techniques for the chemical analysis of such geological samples by measuring elemental composition upon ablation of target areas, often performed on mineral grains at the micron scale [@George:2018; @Yao:2017; @Astbury:2018]. Accurate recording of laser ablation targets is important for traceability, ensuring that collected data can be matched to its target location and any other related data. Annotation of laser ablation targets or Regions Of Interest (ROI) for laser ablated samples can be mapped onto high resolution Scanning Electron Microscope (SEM) images, which are often used for mineral characterisation [@Astbury:2018; @Bonnetti:2020]. Historically, this has been a manual process for many laboratories, thus could lead to inaccurate labelling which impedes traceability, verification of conclusions, and compromises important studies.
 
-# Existing Solutions
+# Statement of Need
+
+TACtool (Targeting And Coordination tool) is a PyQt based desktop application which was designed to enable users to annotate laser ablation analysis locations onto high resolution SEM images. TACtool automatically calculates coordinates of annotations which can be uploaded to laser ablation systems, ensuring efficient and accurate re-coordination of target locations relative to the spatial context resolved in the SEM image. In the latest version, TACtool can also import an SEM image and identity point coordinates and plot these existing positions onto the imported image. These can then be used or added to ready for export. This automatic process additionally saves time, lab resources, and therefore money, when identifying ROI. TACtool also provides optional metadata recording for every ROI which further improves reproducibility and traceability throughout a lab's workflow.
+
+The intuitive and accessible front-end design of TACtool allows users with a lack of computational experience to easily record their annotations. The software can be downloaded as a compiled program, mitigating the need to setup complex development environments. This simplicity continues throughout the workflow within TACtool. Users can import an image file onto a canvas, before clicking directly onto that image to add their ROI and respective annotations. The coordinates and metadata for each ROI can be exported to a plain text Comma-Separated Values (CSV) file, ready to be uploaded directly to a laser ablation system.
+
+# Software Design
 
 There exists tools for the annotation of geological samples and integration with laser ablation systems, however they all have varying functionality, accessibility and ease of use (\autoref{tab:table1}).
 
@@ -43,19 +49,15 @@ There exists tools for the annotation of geological samples and integration with
 |GeoStar μGIS TM [@GeoStar]|Manual and automatic sample point selection. Image importation from any source such as SEM. Integration with ICP-MS systems and automatic instrument control.|Proprietary software package written for use in RESOlution instruments and requires a paid license.|
 |ImageJ (FIJI) [@ImageJ]|Many available plugins and features allow customization of the software to perform a variety of tasks.|Free and open source.|
 |Quartz PCI [@QuartzPCI]|PCI allows for SEM image annotation and measurement of features.|Requires a paid license.|
-|uScope Navigator [@uScopeNavigator]|Capable of image processing such as gamma, contrast and bilateral smoothing filters in addition to ROI annotation. May not be compatible with all laser ablation systems.|Requires a paid license/ comes with their microscope systems.|
+|uScope Navigator [@uScopeNavigator]|Capable of image processing such as gamma, contrast and bilateral smoothing filters in addition to ROI annotation. May not be compatible with all laser ablation systems.|Requires a paid license/comes with their microscope systems.|
[…]
