### 2026-01-28 999d481
Re-write software design section
