### 2026-01-30 #8945 @leorudczenko
Hey @arfon, thank you for the clarification on these new requirements for JOSS. Following this guidance, I have updated our TACtool paper again with all of the required headers, sections, and content. As this included new content, I have reviewed these edits with my co-authors, and we are all happy to continue with the new version.

Is there any information you could provide regarding timescales for the review? It seems like this pre-review process hasn't moved since some reviewers were requested and weren't available.
