### 2026-02-05 #8945 @uellue
> Hello [@uellue](https://github.com/uellue), would you be interested in reviewing this submission?

Hi, this sounds interesting and quite related to some issues we are dealing with in TEM.

I've quickly looked through the paper draft and documentation, and realize I'd have many questions as a reviewer about metadata, stages, coordinate systems, interoperability etc. I've found step-by-step instructions, but not enough to get a high-level understanding of how this works exactly, preconditions, compatibility etc. So please only pick me as a reviewer if you are willing to answer these questions by adding quite a bit more detail and background to your documentation. 🙈 Right now I'm left intrigued, but also a bit puzzled.
