### 2026-03-04 637c32c
Installable Python package with pyproject.toml
From what I can see, tactool really only depends on pyqt5 and numpy.
That makes it trivial to set it up as a Python package with a `pyproject.toml`. All the version pins in
the environment files are really not necessary and just make it hard to set up.

Declaring a script in pyproject.toml makes it super simple to run tactool from the command line.

FIXME: Keywords and authors in pyproject.toml.

Possible next steps:

* Review release mechanism to build wheel and add to release assets
* Publish on PyPI
* Publish on conda-forge
