+++ environments/macos-environment.yml
@@ -1,52 +1,0 @@
-name: tactool-macos
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - blas=2.121=openblas
-  - blas-devel=3.9.0=21_osx64_openblas
-  - bzip2=1.0.8=h1de35cc_0
-  - ca-certificates=2023.12.12=hecd8cb5_0
-  - libblas=3.9.0=21_osx64_openblas
-  - libcblas=3.9.0=21_osx64_openblas
-  - libcxx=16.0.6=hd57cbcb_0
-  - libexpat=2.5.0=hf0c8a7f_1
-  - libffi=3.4.4=hecd8cb5_0
-  - libgfortran=5.0.0=13_2_0_h97931a8_3
-  - libgfortran5=13.2.0=h2873a65_3
-  - liblapack=3.9.0=21_osx64_openblas
-  - liblapacke=3.9.0=21_osx64_openblas
-  - libopenblas=0.3.26=openmp_hfef2a42_0
-  - libsqlite=3.45.1=h92b6c6a_0
-  - libzlib=1.2.13=h8a1eda9_5
-  - llvm-openmp=17.0.6=hb6ac08f_0
-  - ncurses=6.4=hcec6c5f_0
-  - numpy=1.26.4=py311hc43a94b_0
-  - openblas=0.3.26=openmp_h6794695_0
-  - openssl=3.2.1=hd75f5a5_0
-  - pip=24.0=pyhd8ed1ab_0
-  - python=3.11.7=h9f0c242_1_cpython
-  - python_abi=3.11=4_cp311
-  - readline=8.2=hca72f7f_0
-  - setuptools=68.2.2=py311hecd8cb5_0
-  - tk=8.6.13=h1abcd95_1
-  - tzdata=2023d=h04d1e81_0
-  - wheel=0.41.2=py311hecd8cb5_0
-  - xz=5.4.5=h6c40b1e_0
-  - pip:
-      - altgraph==0.17.4
-      - flake8==7.0.0
-      - iniconfig==2.0.0
-      - macholib==1.16.3
-      - mccabe==0.7.0
-      - packaging==23.2
-      - pluggy==1.4.0
-      - pycodestyle==2.11.1
-      - pyflakes==3.2.0
-      - pyinstaller==6.4.0
-      - pyinstaller-hooks-contrib==2024.1
-      - pyqt5==5.15.10
-      - pyqt5-qt5==5.15.12
-      - pyqt5-sip==12.13.0
-      - pytest==8.0.0
-prefix: /usr/local/miniconda/envs/tactool-macos
