### 2026-03-04 #10132 @uellue
All in all it looks like a simple, but very handy tool with a clearly defined use case: Ergonomic coordinate labeling for further downstream processing. However, a few points from the review checklist are still open for me. Most importantly, I'd need more and better high-level documentation to understand what TACtool does, exactly, how to apply it in practice and also use it beyond a very specific laser ablation tool. Such wider application is advertised in the "Research Impact Statement", meaning the documentation should match this claim.

## Installation

* Developer installation can be made much simpler by turning it into a proper Python package. The provided installation method with very specific environment files didn't work for me because Anaconda repositories are blocked for my organization and we have to use `conda-forge` only. I've opened a PR with changes that made developer installation very easy for me and follows the pattern of many Python-based projects: https://github.com/BritishGeologicalSurvey/tactool/pull/17

## State of the field

* Which tools are supported / reported to work / still TODO / ...? Right now it seems only one specific tool by ESI is supported? It looks like the role and functionality of labeling software is closely intertwined with the workflows and user interfaces of laser ablation tools, so the two should be put in relation here.

## Documentation, functionality

* The referencing process between SEM and laser ablation coordinate systems should be described more clearly. In the GUI I found an option to set "reference points", but I couldn't easily work out what I'd have to do with a real-world SEM image. Also, I have a number of additional questions:
  - Does TACtool read any kind of metadata, e.g. scale and offset, from the SEM images or are they just plain images with pixel coordinates?
  - Specifically, do the reference points have to be included in the same image as the intended measurements, or can the software relate them through embedded metadata in the SEM images?
  - What SEM image formats are supported? Anything specific from ISO/DIS 20171, which seems relevant for SEM?
  - The role of "scale" is completely unclear to me. TACtool asks for me to set it, but how and why was not easy to figure out.
  - How exactly does the laser ablation tool deal with the referencing? In the very detailed "ESI import" instructions it seems that the user finds and marks their reference points using the GUI of the tool? Do all tools work the same or are there differences? That should be described in a generic high-level workflow description.
  - Is it possible to output coordinates in the CSV file in a calibrated coordinate system, e.g. by assigning coordinates to the reference points, or are the coordinates always calibrated and transformed by a downstream tool, like it seems in the ESI example? The paper hints at generating lists of _calibrated_ coordinates, so this should be clarified.
  - How should one approach reference points in practice? Are alignment marks included on the mounts, or does one pick specific features on the specimen itself? Or does one add markers somehow during specimen preparation or imaging? Any practical guidance will probably be valuable for novice users!
  - What SEM magnifications are a good choice? It seems that the SEM can achieve a much higher spatial resolution than optical images from the laser ablation tool?
  - How can a user check if the alignment is correct? E.g. look for the "impact craters" in SEM images after ablation? I'm not an expert for this method at all, but since you mention reproducibility and reliability, would it make sense to collect another set of SEM images post ablation routinely to verify the correct locations were hit, or verify the locations through some other means?
* The schema for CSV files generated by TACtool resp. accepted by various ablation tools should be described or referenced. Right now one has to guess. In particular, units are important and not always obvious! Is it the same for all tools, or is the input vendor-specific? Is a standard or quasi-standard established and documented somewhere?
*  Is it possible to specify/load schemas for different tools? It would be great if the output was configurable, that would make it much more useful for wider application! We have very similar issues in TEM -- relating coordinates from different analysis methods.
* The `instructions.md` should include more high-level workflow descriptions. Right now it contains a lot of info that is kind of obvious ("material: The material assigned to the analysis point", detailed instructions on how to use a common file selection dialogue, ...), but not what role that info plays in the larger picture. E.g., is "material" interpreted by the ablation tool, and how does it influence results? What role does the colour play? ...

Since I don't have access to laser ablation tools, I can't test the functionality completely. I'd check off the functionality and documentation items once I can understand from the documentation how the referencing process works for the targeted laser ablation tools, what role TACtool and the generated CSV files play, and the functionality is described in such a way that I can gauge if and how I can apply TACtool for techniques that I am more familiar with, such as TEM-related workflows.
