### 2026-03-10 #10132 @leorudczenko
Hi there @uellue, thank you for the very prompt review of our software and paper TACtool! To fully answer your questions, I need to speak to my co-authors as I do not have all of the answers alone. In the meantime I will review your merge request to use a `pyproject.toml` to replace the pinned conda dependencies.
