### 2026-03-10 c2bbd95
Apply pyproject to github workflow as test
