+++ .github/workflows/test.yml
@@ -19,9 +19,15 @@ jobs:
     - name: Activate Anaconda Environment
       uses: conda-incubator/setup-miniconda@v2
       with:
-        environment-file: environments/${{ matrix.os }}-environment.yml
-        python-version: 3.11
+        python-version: 3.14
         auto-activate-base: false
+
+    - name: Build Environment
+      shell: bash -l {0}
+      run: |
+        conda create -n tactool python=3.14
+        conda activate tactool
+        pip install -e .[test]
 
     - name: Lint
       shell: bash -l {0}
