+++ .github/workflows/main.yml
@@ -13,9 +13,9 @@ jobs:
           - os: windows
             app_file_ext: "exe"
             asset_file_ext: "exe"
-          # - os: macos
-          #   app_file_ext: "app"
-          #   asset_file_ext: "zip"
+          - os: macos
+            app_file_ext: "app"
+            asset_file_ext: "zip"
     runs-on: ${{ matrix.os }}-latest
     steps:
     - uses: actions/checkout@v3
@@ -32,7 +32,7 @@ jobs:
       run: |
         conda create -n tactool python=3.13
         conda activate tactool
-        pip install -e .[pyinstaller]
+        pip install -e ".[pyinstaller]"
 
     - name: Build with PyInstaller
       shell: bash -l {0}
