### 2026-03-10 f7bfe93
Set global shell
