### 2026-03-10 4dfd253
Manually activate env each step
