### 2026-03-10 0ef62fb
Make environment persistent between steps
