+++ .github/workflows/test.yml
@@ -22,6 +22,7 @@ jobs:
     - name: Activate Anaconda Environment
       uses: conda-incubator/setup-miniconda@v2
       with:
+        miniconda-version: latest
         activate-environment: tactool
         auto-activate-base: false
 
