### 2026-03-10 97c9f3d
Run on changes to pyproject
