+++ pyproject.toml
@@ -9,7 +9,7 @@ include = ["tactool"]
 [project]
 name = "tactool"
 description = "TACtool: A targeting and co-ordination tool for spatial sampling and data traceability applied to scanning electron microscopy and laser ablation ICP-MS"
-version = "1.4.0.dev0"
+version = "1.3.1.dev0"
 license = "GPL-3.0-only"
 license-files = ["LICENSE"]
 keywords = ["FIXME"]
