### 2026-03-10 6e99b09
Add quotes to pip command installing optional dependencies
