### 2026-03-10 f754b56
Update GitHub actions workflows to use pyproject installation
