### 2026-03-31 fdd9d1c
Update pyproject.toml metadata
