### 2026-04-07 a4b24ec
Add more detail to high level workflow description
