### 2026-04-07 7167080
Add more detail to instructions
