### 2026-04-07 #10132 @leorudczenko
Hi @uellue, I have some answers to your questions below. Please let us know if we haven't answered them in enough detail. I have opened a merge request on the TACtool GitHub repository to update the documentation to better reflect these answers and would appreciate your feedback on these changes: https://github.com/BritishGeologicalSurvey/tactool/pull/18

## Installation
- Developer installation can be made much simpler by turning it into a proper Python package. The provided installation method with very specific environment files didn't work for me because Anaconda repositories are blocked for my organization and we have to use conda-forge only. I've opened a PR with changes that made developer installation very easy for me and follows the pattern of many Python-based projects: https://github.com/BritishGeologicalSurvey/tactool/pull/17
    - **Answer**: _I have reviewed and merged your merge request to implement a `pyproject.toml` based development environment instead of the strict `conda` environment YAML structure. Thank you very much for your work on this!_

## State of the field
- Which tools are supported / reported to work / still TODO / ...? Right now it seems only one specific tool by ESI is supported? It looks like the role and functionality of labeling software is closely intertwined with the workflows and user interfaces of laser ablation tools, so the two should be put in relation here.
    - **Answer**: _TACtool is not designed to support other software; it is a standalone package which can export images and co-ordinates to be used by other software. TACtool is use and platform agnostic, it can be used just as well for any image that requires location co-ordination. Our preferred application is for laser ablation analysis but this will work just as well for other high-spatial resolution analytical techniques (SIMS, SEM, EPMA) but is also scale independent so can be used for images of larger objects (e.g. a house) that require positional referencing (of objects inside)._

## Documentation, functionality
- The referencing process between SEM and laser ablation coordinate systems should be described more clearly. In the GUI I found an option to set "reference points", but I couldn't easily work out what I'd have to do with a real-world SEM image. Also, I have a number of additional questions:
    - Does TACtool read any kind of metadata, e.g. scale and offset, from the SEM images or are they just plain images with pixel coordinates?
        - **Answer**: _TACtool does not read any metadata from any SEM image, they are just plain images with pixel coordinates._

    - Specifically, do the reference points have to be included in the same image as the intended measurements, or can the software relate them through embedded metadata in the SEM images?
        - **Answer**: _The reference points have to be included in the same image as the intended measurements._

    - What SEM image formats are supported? Anything specific from ISO/DIS 20171, which seems relevant for SEM?
        - **Answer**: _Any image is passed straight to a `PyQt5.QtGui.QPixmap` object, which accepts the following: BMP, CUR, GIF, ICO, JFIF, JPEG, JPG, PBM, PGM, PNG, PPM, SVG, TIFF, XBM, XPM._

    - The role of "scale" is completely unclear to me. TACtool asks for me to set it, but how and why was not easy to figure out.
        - **Answer**: _The scale tool scales the size of the analysis spot set on the image relative to the objects and scale bar in the image (or a relative input measurement (e.g. 200um for the length of a large zircon) if a scale bar is absent)._

    - How exactly does the laser ablation tool deal with the referencing? In the very detailed "ESI import" instructions it seems that the user finds and marks their reference points using the GUI of the tool? Do all tools work the same or are there differences? That should be described in a generic high-level workflow description.
        - **Answer**: _Most re-coordination systems work by using a translation matrix on the two sets of co-ordinates for the same location positions for the two systems being co-ordinated, in this case TACtool and a laser ablation system._

    - Is it possible to output coordinates in the CSV file in a calibrated coordinate system, e.g. by assigning coordinates to the reference points, or are the coordinates always calibrated and transformed by a downstream tool, like it seems in the ESI example? The paper hints at generating lists of calibrated coordinates, so this should be clarified.
        - **Answer**: _TACtool assigns co-ordinates to the reference points. The import tool of the next system then uses these along with its own co-ordinates for these reference points, through the use of a translation matrix, to cross-coordinate the two data sets._

    - How should one approach reference points in practice? Are alignment marks included on the mounts, or does one pick specific features on the specimen itself? Or does one add markers somehow during specimen preparation or imaging? Any practical guidance will probably be valuable for novice users!
        - **Answer**: _Any of the above can be done. In practice, in the absence of predetermined markers, it is often easier to use specific features on the specimen itself which can be accurately identified in the next co-ordination system. The accuracy of the re-coordination depends on how accurately the reference points can be identified between the two systems._

    - What SEM magnifications are a good choice? It seems that the SEM can achieve a much higher spatial resolution than optical images from the laser ablation tool?
        - **Answer**: _Yes the SEM has a higher spatial resolution than the optical images in the laser ablation tool. I do not think that the SEM magnification is significant here._

    - How can a user check if the alignment is correct? E.g. look for the "impact craters" in SEM images after ablation? I'm not an expert for this method at all, but since you mention reproducibility and reliability, would it make sense to collect another set of SEM images post ablation routinely to verify the correct locations were hit, or verify the locations through some other means?
        - **Answer**: _The TACtool image with marked location points, can be exported along with the co-ordinates. This image can be imported into the laser ablation system as a layer and its opacity reduced so that the outline of the image and spot location can be seen relative to the physical sample position and the analysis spot set in the laser system. The two can thereby be seen to overlap. With accurate cross-coordination this can be a perfect overlap. However, TACtool is a standalone system; it generates a set of co-ordinates for locations within an image where there were none. The accuracy of alignment is therefore down to the re-coordination algorithm of the importing system, which TACtool has no influence over._

- The schema for CSV files generated by TACtool resp. accepted by various ablation tools should be described or referenced. Right now one has to guess. In particular, units are important and not always obvious! Is it the same for all tools, or is the input vendor-specific? Is a standard or quasi-standard established and documented somewhere?
    - **Answer**: _I am not 100% sure what the question is here, but here is my best answer. The schema for the CSV files includes the following columns: Name, Type, X, Y, diameter, scale, colour, mount\_name, material, notes. Only the columns which start with a capital letter are used by the laser ablation tool (Name, Type, X, Y); the rest of the metadata passes through the system and is retained for traceability. For a CSV file to be compatible with TACtool, it only requires the first 4 capitalised columns. Regarding the units we use, these are in the GUI and in the instructions file here: https://github.com/BritishGeologicalSurvey/tactool/blob/main/docs/instructions.md_

- Is it possible to specify/load schemas for different tools? It would be great if the output was configurable, that would make it much more useful for wider application! We have very similar issues in TEM -- relating coordinates from different analysis methods.
    - **Answer**: _It is not possible to use a different schema currently. However as mentioned above, TACtool only requires the columns (Name, Type, X, Y) in a CSV file. TACtool is designed to be system agnostic so can be used with any platform requiring co-ordination._

- The instructions.md should include more high-level workflow descriptions. Right now it contains a lot of info that is kind of obvious ("material: The material assigned to the analysis point", detailed instructions on how to use a common file selection dialogue, ...), but not what role that info plays in the larger picture. E.g., is "material" interpreted by the ablation tool, and how does it influence results? What role does the colour play?
    - **Answer**: _These labels are simply metadata so that the analyst knows the material (e.g. mineral) that was ablated at the time. It isn’t used in the results, nor is the colour, that’s just to make it easier to see on different image backgrounds, or perhaps to set different groupings of points. Its just functionality and metadata, allowing understanding and traceability of practice at a later date._
