### 2026-04-30 #10132 @uellue
Thank you for the update and answers! Reading through the paper with this information in mind, I'd suggest the following changes:

* Line 25 "TACtool automatically calculates coordinates of annotations which can be
uploaded to laser ablation systems, ..." maybe better say "TACtool records pixel coordinates of annotations which can be
uploaded to laser ablation systems..." since TACtool seems to work in pixel coordinates and the coordinate transformation is actually performed by the laser ablation tool? I think that was the statement that made me expect some form of transformation being performed by TACtool. The same statement should also be adjusted in the README.
* Line 39+: "..., ready to be uploaded directly to a laser ablation system": It seems that it is only tested with ESI? My question about compatibility with tools from other vendors and the supported data schemas wasn't really answered, so one could limit this statement e.g. to "ready to be uploaded directly to an **ESI** laser ablation system or converted to another format.". Same for line 58+ " TACtool can then export the ROI and associated metadata to a simple CSV file, in a format which is recognised by **ESI** laser ablation systems. If it was positively confirmed that the exported CSV files also work with other specific laser ablation tools, one could include a table of tested and supported tools in the documentation and change the statement to a "compatible laser ablation system".
