### 2026-04-30 cb90faf
Specify pixel coordinates and ESI laser ablation systems
