+++ paper/paper.md
@@ -36,9 +36,9 @@ The precise analysis of geological samples is crucial for understanding the hist
 
 # Statement of Need
 
-TACtool (Targeting And Coordination tool) is a PyQt based desktop application which was designed to enable users to annotate laser ablation analysis locations onto high resolution SEM images. TACtool automatically calculates coordinates of annotations which can be uploaded to laser ablation systems, ensuring efficient and accurate re-coordination of target locations relative to the spatial context resolved in the SEM image. In the latest version, TACtool can also import an SEM image and identity point coordinates and plot these existing positions onto the imported image. These can then be used or added to ready for export. This automatic process additionally saves time, lab resources, and therefore money, when identifying ROI. TACtool also provides optional metadata recording for every ROI which further improves reproducibility and traceability throughout a lab's workflow.
+TACtool (Targeting And Coordination tool) is a PyQt based desktop application which was designed to enable users to annotate laser ablation analysis locations onto high resolution SEM images. TACtool records pixel coordinates of annotations which can be uploaded to laser ablation systems, ensuring efficient and accurate re-coordination of target locations relative to the spatial context resolved in the SEM image. In the latest version, TACtool can also import an SEM image and identity point coordinates and plot these existing positions onto the imported image. These can then be used or added to ready for export. This automatic process additionally saves time, lab resources, and therefore money, when identifying ROI. TACtool also provides optional metadata recording for every ROI which further improves reproducibility and traceability throughout a lab's workflow.
 
-The intuitive and accessible front-end design of TACtool allows users with a lack of computational experience to easily record their annotations. The software can be downloaded as a compiled program, mitigating the need to setup complex development environments. This simplicity continues throughout the workflow within TACtool. Users can import an image file onto a canvas, before clicking directly onto that image to add their ROI and respective annotations. The coordinates and metadata for each ROI can be exported to a plain text Comma-Separated Values (CSV) file, ready to be uploaded directly to a laser ablation system.
+The intuitive and accessible front-end design of TACtool allows users with a lack of computational experience to easily record their annotations. The software can be downloaded as a compiled program, mitigating the need to setup complex development environments. This simplicity continues throughout the workflow within TACtool. Users can import an image file onto a canvas, before clicking directly onto that image to add their ROI and respective annotations. The coordinates and metadata for each ROI can be exported to a plain text Comma-Separated Values (CSV) file, ready to be uploaded directly to an ESI laser ablation system.
 
 # State of the Field
 
@@ -59,7 +59,7 @@ Many of the existing paid software solutions are efficient, but the fact that th
 
 # Software Design
 
-When designing and building TACtool, there were 3 main goals we aimed for. Firstly, we wanted to produce a front-end application that is lightweight and easy to use, whilst still providing useful traceability. To fulfil this, we have included a very limited number of settings which allow some customisation and sample metadata entry. TACtool can then export the ROI and associated metadata to a simple CSV file, in a format which is recognised by laser ablation systems. Secondly, Python and the PyQt library was chosen to develop the code, as Python is widely used in the scientific community, and the underlying Qt library is widely used to build graphical user interfaces. Using what is already widely known allows others in the community to more easily contribute to the TACtool code. Thirdly, TACtool was to be distributed as a single executable, with no setup required by the end user. We used GitHub Actions to achieve this by setting up an automated pipeline which takes the latest version of the code and compiles it. We are then able to provide a single download button in the README document, making it as simple as possible for end users to get started.
[…]
