### 2026-04-30 dbeeda3
Specify pixel coordinates and ESI laser ablation systems
