### 2026-04-30 #10132 @leorudczenko
@uellue I agree with your suggestions regarding pixel coordinate wording and specifying ESI laser ablation systems. I've made these changes now to both the README and the paper itself. Thank you for taking the time to respond!
