### 2026-04-30 #10132 @uellue
Great! I've checked the remaining items based on the latest proof and anticipating merging of https://github.com/BritishGeologicalSurvey/tactool/pull/18 for the docs changes. That concludes my review!
