### 2026-04-30 #10132 @uellue
...ah, about the "AI use", I've used ChatGPT for a background search on coordinate annotation software to see if anything important was missing from the State of the Field.
