+++ .github/workflows/test.yml
@@ -17,10 +17,10 @@ jobs:
       run:
         shell: bash -l {0}
     steps:
-    - uses: actions/checkout@v3
+    - uses: actions/checkout@a12a3943b4bdde767164f792f33f40b04645d846  # v3
 
     - name: Activate Anaconda Environment
-      uses: conda-incubator/setup-miniconda@v2
+      uses: conda-incubator/setup-miniconda@11ae174708d1ae769b9457e7d6ed64d606c99af1  # v2
       with:
         miniconda-version: latest
         activate-environment: tactool
