### 2026-04-30 f4c5175
Use GitHub Action hashes instead of version numbers for paper
