+++ .github/workflows/paper.yml
@@ -12,7 +12,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@1e31de5234b9f8995739874a8ce0492dc87873e2  # v4
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -20,7 +20,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: paper/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v4
+        uses: actions/upload-artifact@c7d193f32edcb7bfad88892161225aeda64e9392  # v4
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
