### 2026-05-01 f73cd85
Add comment about using scale to ensure diameter is applied correctly
